+++ core/src/test/java/org/eclipse/hono/util/CredentialsObjectTest.java
@@ -13,8 +13,6 @@
 
 package org.eclipse.hono.util;
 
-import static org.junit.Assert.*;
-
 import org.junit.Test;
 
 import io.vertx.core.json.JsonObject;
@@ -30,10 +28,10 @@ public class CredentialsObjectTest {
      * Verifies that credentials that do not contain any secrets are
      * detected as invalid.
      */
-    @Test
-    public void testHasValidCredentialsDetectsMissingSecrets() {
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsMissingSecrets() {
         final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
-        assertFalse(creds.hasValidSecrets());
+        creds.checkSecrets();
     }
 
     /**
@@ -41,9 +39,73 @@ public class CredentialsObjectTest {
      * considered valid.
      */
     @Test
-    public void testHasValidCredentialsAcceptsEmptySecrets() {
+    public void testCheckSecretsAcceptsEmptySecrets() {
         final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
         creds.addSecret(new JsonObject());
-        assertTrue(creds.hasValidSecrets());
+        creds.checkSecrets();
+    }
+
+    /**
+     * Verifies that credentials that contain a malformed not-before value are
+     * detected as invalid.
+     */
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsMalformedNotBefore() {
+        final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
+        creds.setType(CredentialsConstants.SECRETS_TYPE_PRESHARED_KEY);
+        creds.addSecret(new JsonObject()
+                .put(CredentialsConstants.FIELD_SECRETS_NOT_BEFORE, "malformed"));
+        creds.checkSecrets();
+    }
+
+    /**
+     * Verifies that credentials that contain a malformed not-before value are
+     * detected as invalid.
+     */
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsMalformedNotAfter() {
+        final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
+        creds.setType(CredentialsConstants.SECRETS_TYPE_PRESHARED_KEY);
+        creds.addSecret(new JsonObject()
+                .put(CredentialsConstants.FIELD_SECRETS_NOT_AFTER, "malformed"));
+        creds.checkSecrets();
+    }
+
+    /**
+     * Verifies that credentials that contain inconsistent values for not-before
+     * and not-after are detected as invalid.
+     */
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsInconsistentValidityPeriod() {
+        final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
+        creds.setType(CredentialsConstants.SECRETS_TYPE_PRESHARED_KEY);
+        creds.addSecret(new JsonObject()
+                .put(CredentialsConstants.FIELD_SECRETS_NOT_BEFORE, "2018-10-10T00:00:00+00:00")
+                .put(CredentialsConstants.FIELD_SECRETS_NOT_AFTER, "2018-10-01T00:00:00+00:00"));
+        creds.checkSecrets();
+    }
+
+    /**
+     * Verifies that hashed-password credentials that do not contain the hash function name are
+     * detected as invalid.
+     */
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsMissingHashFunction() {
+        final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
+        creds.setType(CredentialsConstants.SECRETS_TYPE_HASHED_PASSWORD);
+        creds.addSecret(new JsonObject().put(CredentialsConstants.FIELD_SECRETS_PWD_HASH, "hash"));
+        creds.checkSecrets();
+    }
+
+    /**
+     * Verifies that hashed-password credentials that do not contain the hash value are
+     * detected as invalid.
+     */
+    @Test(expected = IllegalStateException.class)
+    public void testCheckSecretsDetectsMissingHashValue() {
+        final CredentialsObject creds = new CredentialsObject("tenant", "device", "x509-cert");
+        creds.setType(CredentialsConstants.SECRETS_TYPE_HASHED_PASSWORD);
+        creds.addSecret(new JsonObject().put(CredentialsConstants.FIELD_SECRETS_HASH_FUNCTION, CredentialsConstants.HASH_FUNCTION_SHA256));
+        creds.checkSecrets();
     }
 }