+++ services/device-registry/src/test/java/org/eclipse/hono/deviceregistry/FileBasedCredentialsServiceTest.java
@@ -23,6 +23,7 @@ import java.net.HttpURLConnection;
 import java.nio.charset.StandardCharsets;
 
 import org.eclipse.hono.service.credentials.CompleteCredentialsService;
+import org.eclipse.hono.util.ClearTextPassword;
 import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.CredentialsConstants;
 import org.eclipse.hono.util.CredentialsObject;
@@ -237,7 +238,11 @@ public class FileBasedCredentialsServiceTest {
         when(fileSystem.existsBlocking(FILE_NAME)).thenReturn(Boolean.TRUE);
         final Async add = ctx.async(2);
         final CredentialsObject hashedPassword = CredentialsObject.fromHashedPassword(
-                "4700", "bumlux", "secret", CredentialsConstants.HASH_FUNCTION_SHA512, null, null, null);
+                "4700",
+                "bumlux",
+                ClearTextPassword.encode(CredentialsConstants.HASH_FUNCTION_SHA512, null, "secret"),
+                CredentialsConstants.HASH_FUNCTION_SHA512,
+                null, null, null);
         final CredentialsObject psk = CredentialsObject.fromPresharedKey(
                 "4711", "sensor1", "sharedkey".getBytes(StandardCharsets.UTF_8), null, null);
         svc.add(